requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
psycopg2-binary
pymysql